feat: semantic URL routing — redirects, URL alignment, slug stability
1. 301 redirects from /wiki/E{N} → semantic directory URLs
   /wiki/E42 now permanently redirects to /organizations/anthropic
   for entity types with dedicated directory pages.

2. llms.txt uses semantic URLs
   generate-llm-files.mjs now emits /organizations/anthropic instead
   of /wiki/E42 for directory-backed entities.

3. Publications promoted to /publications/[id]
   Moved from /sources/publications/[id] to top-level route.
   Old URL permanently redirects to new one.

4. Resources promoted to /resources/[id]
   Moved from /source/[id] to top-level route.
   All internal links updated. Old URL permanently redirects.

5. Slug stability via previousSlugs
   Entities can declare previousSlugs in YAML. Directory routes
   check aliases and permanently redirect old slugs to current ones.
   Supported in KB types, loader, serializer, and all 5 directory routes.

Closes #2046
