+++ apps/web/src/app/publications/page.tsx
@@ -1,0 +1,5 @@
+import { redirect } from "next/navigation";
+
+export default function PublicationsPage() {
+  redirect("/wiki/E1044");
+}
