+++ apps/web/src/app/resources/page.tsx
@@ -1,0 +1,5 @@
+import { permanentRedirect } from "next/navigation";
+
+export default function ResourcesListingPage() {
+  permanentRedirect("/wiki/E1043");
+}
